Revert "Optimized the calendar so that it filters from a logged in user and gets their pets and notifications corresponding rather than just retrieving all the notifications in the db regardless of user. This allows for filtering via petName but also notification type (currently unfinished)."
This reverts commit 2bc680dc880a90fe860eb741a092ee4aad032112.
