+++ prisma/schema.prisma
@@ -63,9 +63,9 @@ model Notification {
   frequencyQuantity Int?
   frequencyUnit FrequencyUnit? 
   day String[] 
-  time String[] 
+  time String[]
   endDate String[] 
-  repeating String[] 
+  repeating Repeating[] 
   createdAt DateTime @default(now())
   imageSrc String?
   completed Boolean @default(false)
