--- src/typings/user.d.ts
@@ -1,10 +1,0 @@
-type User = {
-  id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  password: string;
-  username: string;
-};
-
-export { User };
