+++ tests/api/Users/user.rest
@@ -10,9 +10,6 @@ GET http://localhost:3001/users/clte5s2lp0000st8dcrhqf8jt
 ## Request 
 GET http://localhost:3001/users/username/johnsmith1
 
-### Description: Retrieve pet notification via pet's full name
-## Request 
-GET http://localhost:3001/users/petData/johnsmith1/Toasty/Smithy
 
 ### Description: Create user 
 ## Request
